Add unauthorized check and page for notifying user. Plan to also hide or disable CRUD buttons if user can't use them.
