+++ QAPPBuilder/qapp_builder/templates/401.html
@@ -1,0 +1,9 @@
+{% extends "main/base.html" %}
+
+{% block content %}
+    <div class="grid-container">
+        <h1>Unauthorized</h1>
+        <p>You don't have permission to modify the selected object.</p>
+        <button class="usa-button" onclick="window.history.back()">Go Back</button>
+    </div>
+{% endblock %}
